refactor(cli): extract session interaction helpers
