+++ packages/agent-cli/agent-cli.cabal
@@ -56,6 +56,7 @@ library
                     , Agent.CLI.Session.Attachments
                     , Agent.CLI.Session.Choices
                     , Agent.CLI.Session.History
+                    , Agent.CLI.Session.Interaction
                     , Agent.CLI.Session.Selection
                     , Agent.CLI.Session.Runtime.Types
                     , Agent.CLI.SessionAdmin
